Use Jupyter name instead of IPython across code base
Jupyter is the future so we shouldn't start off in the past.
Refs #21251
